This fixes #2799: With `draw_empty=false`, the margin container produced negative widths and heights for its inner widget. The first commit in here fixes that.

The second commit adds assertions to catch negative widget sizes early. Any code that gets broken by this change was most likely already broken before. And due to the many `pcall`s in the widget layout code, this error will be contained and will not "break everything".

The last commit fixes the `empty_clip` function to actually work. However, for reasons that I do not really understand, this does not have much of an effect on `test-benchmark.lua`.
Master:
```
   create&draw wibox: 0.0140362  sec/iter ( 73 iters, 1.6 sec for benchmark)
    update textclock: 2.68101e-06 sec/iter (428868 iters, 1.383 sec for benchmark)
  relayout textclock: 1.70189e-06 sec/iter (808534 iters, 1.582 sec for benchmark)
    redraw textclock: 8.05295e-07 sec/iter (1260045 iters, 3.085 sec for benchmark)
          tag switch: 0.0054088  sec/iter (185 iters, 1.674 sec for benchmark)
```
With this change:
```
   create&draw wibox: 0.0140465  sec/iter ( 73 iters, 1.59 sec for benchmark)
    update textclock: 2.68279e-06 sec/iter (428028 iters, 1.343 sec for benchmark)
  relayout textclock: 1.73115e-06 sec/iter (797147 iters, 1.631 sec for benchmark)
    redraw textclock: 7.87591e-07 sec/iter (1289767 iters, 3.115 sec for benchmark)
          tag switch: 0.00541387 sec/iter (186 iters, 2.68 sec for benchmark)
```